Abstracted out cron command; added scheduler cron command; moved command names to sdss:drupal:cron.
