+++ blt/src/Blt/Plugin/Commands/GryphonCommands.php
@@ -3,8 +3,6 @@
 namespace Gryphon\Blt\Plugin\Commands;
 
 use Acquia\Blt\Robo\BltTasks;
-use Acquia\Blt\Robo\Common\EnvironmentDetector;
-use GuzzleHttp\Client;
 
 /**
  * Class GryphonCommands.
@@ -34,43 +32,6 @@ class GryphonCommands extends BltTasks {
     }
 
     $commands->run();
-  }
-
-  /**
-   * Run the Drupal cron on all sites.
-   *
-   * @command drupal:cron
-   */
-  public function cron() {
-    $this->config->set('drush.alias', '');
-    $failed = [];
-    foreach ($this->getConfigValue('multisites') as $multisite) {
-      try {
-        $this->say("Running Cron on <comment>$multisite</comment>...");
-        $this->switchSiteContext($multisite);
-
-        $task = $this->taskDrush()
-          ->drush("cron")
-          ->run();
-        if (!$task->wasSuccessful()) {
-          $failed[] = $multisite;
-          $this->say("Cron on <comment>$multisite</comment> failed.");
-          continue;
-        }
-        $this->say("Cron on <comment>$multisite</comment> completed successfully.");
-      }
-      catch (\Exception $e) {
-        $this->say("Unable to run cron on <comment>$multisite</comment>: " . $e->getMessage());
-        $failed[] = $multisite;
-        continue;
-      }
-    }
-
-    if ($failed) {
-      $this->say("Cron failed on the following sites: " . implode(', ', $failed));
-    }
-
-    $this->say("Cron completed successfully on all sites.");
   }
 
   /**
